\name{RStorm.env}
\alias{RStorm.env}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Environment used by RStorm for internal storage of objects.
}

\description{
Environment used by RStorm for internal storage of objects. These objects facilitate the functioning of the Stream and store the emitted Tuples in between Bolts.
}

\author{
	Maurits Kaptein
}

\seealso{
	See Also: \code{\link{RStorm}}
}

\keyword{ utilities }
\keyword{ programming }
