\name{sentences}
     \docType{data}
     \alias{sentences}
     \title{Sentences of the first section of the paper by Student Introducing the T-Test.}
     \description{
       This dataset gives a number of sentences -- the first sentences from the article ``The probable error of a mean'' by Student.
     }
     \usage{rivers}
     \format{A data.frame containing two columns, one with an ID (number) for each sentence, and one with the sentences itself.}
     \source{The probable error of a mean, Biometrica, 1908}
     \references{
       Student (1908) \emph{The probable error of a mean}. Biometrica, 6, 1, 1--25.
     }
     \keyword{datasets}